+++ changes.md
@@ -1,1 +1,1 @@
-* none
+* elevation: arrows on track if distanceMarkers = true (https://github.com/Raruto/leaflet-elevation/issues/244)